+++ npm/ng-packs/packages/oauth/src/lib/utils/auth-utils.ts
@@ -6,28 +6,9 @@ import {
   ConfigStateService,
   LoginParams,
   PipeToLoginFn,
-  AbpLocalStorageService,
 } from '@abp/ng.core';
 import { OAuthService } from 'angular-oauth2-oidc';
-
-const rememberMe = 'remember_me';
-
-export class RememberMeService {
-  constructor(private injector: Injector) { }
-  localStorageService = this.injector.get(AbpLocalStorageService);
-
-  setRememberMe(remember: boolean) {
-    this.localStorageService.setItem(rememberMe, JSON.stringify(remember));
-  }
-
-  removeRememberMe() {
-    this.localStorageService.removeItem(rememberMe);
-  }
-
-  getRememberMe() {
-    return this.localStorageService.getItem(rememberMe);
-  }
-}
+import { RememberMeService } from '../services/remember-me.service';
 
 export const pipeToLogin: PipeToLoginFn = function (
   params: Pick<LoginParams, 'redirectUrl' | 'rememberMe'>,
@@ -35,11 +16,11 @@ export const pipeToLogin: PipeToLoginFn = function (
 ) {
   const configState = injector.get(ConfigStateService);
   const router = injector.get(Router);
-  const rememberMeService = new RememberMeService(injector);
+  const rememberMeService = injector.get(RememberMeService);
   return pipe(
     switchMap(() => configState.refreshAppState()),
     tap(() => {
-      rememberMeService.setRememberMe(params.rememberMe);
+      rememberMeService.set(params.rememberMe);
       if (params.redirectUrl) router.navigate([params.redirectUrl]);
     }),
   );
